+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.3.0 - 2017-01-14
+- using `onDidChangeConfiguration` API to detect configuration change and make the user settings take effect
+- adding command `Toggle highlight` to enable/disable the highlight
+- adding a configuration section `todohighlight.isEnable` to enable/disable the highlight
+
 ## 0.2.1 - 2017-01-06
 - fix #5